requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.22.9
botocore==1.25.9
distlib==0.3.4
Django==4.0.4
djangorestframework==3.13.1
filelock==3.6.0
jmespath==1.0.0
platformdirs==2.5.2
pysqlite3==0.4.7
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
virtualenv==20.14.1
PyJWT~=2.4.0
cryptography~=37.0.2